requirements.txt
discord
google-generativeai
wikipedia
requests
